Do not shrink commit index when a snapshot is installed.
The original goal was to avoid the situation where commit index is
larger than log size. However the situation could also happen when
the leader does not send the entire log but moved commit index.

After this change we could guarantee that commit index never moves
back, and avoid the the possibility of delivering a commit at the
same index twice with different values.

The draw back is that sometimes a snapshot is not taken at the
optimal moment.