Do not shrink log beyond log start.
We never deliver a log entry to the application if it is not
committed. In addition, we never snapshot a log entry if it is not
accepted by the application.

Combine those two, we never snapshot a log entry if it is not
comitted. That is to say, the snapshot only contains committed
entries. Furthermore, a comitted entry can never be rolled back.
Thus we never need to rollback any snapshot, which is the log start.

To achive that effect, the commit index is sent back to the leader
so that the leader won't send logs beyond that point. Append entries
and install snapshot requests are still accepted if we can check the
previous log entry, even if they are before (<) the commit index. If
the entry cannot be verified (i.e. before log start), we respond with
a failure plus the commit index.

After this change, we can be sure that commit index is always within
the current log range, with the exception of newly restored logs. In
that case the commit index is simply set to the log start.

Coincidentally, a lot of changes made in the previous commit were
reverted. That does not mean they are wrong. The cases were greatly
simplified after strict log rules are imposed.