Remove a trigger condition of log_grow.
The snapshot index never goes backwards either, just like log start.
So it is always larger or equal to (>=) log start.

The trigger condition always holds. That does not mean we should
always trigger. On the contrary, it means the target situation the
condition is trying to address is impossible.