Create one log file for each test.
This is done by storing the logger in each thread. Each newly created
thread should inherit the logger from the parent thread.

A feature integration-test is created for conditionally enabling. Note that
`#[cfg(test)]` does not work in integration tests, because they are
considered external binaries.

Create a feature to make sure test is run.