Create an abstract layer on top of RPC interfaces for remote peers.
A new trait RemoteRuaft is added to represent a Raft peer. The RPC layer
should provide an implementation that sends messages to the peer and
conveys the response back.

With RemoteRuaft, Ruaft can be run on different RPC layers easily.

An implementation of RemoteRuaft must be static because it will be sent
between threads.