Reduce the frequency of heartbeats when there are pending commits.
If there are pending commits which are about to be committed, we can
rely on new commits to prove authority.

This change reduced CPU usage for about 10%, can support a 1.5x rate
of read only requests when the network is good. It does not affect
capacity when the network is unreliable. Overall this is a pure
improvement in all aspects.