Replace Arc<Runtime> with a custom thread pool holder.
This change replaces an indirect ownership (Arc<Runtime>) with a direct
ownership (ThreadPoolHolder owns a Handler), and thus improves efficiency.

A Runtime and a Handle are essentially the same thing under the hood. The
difference is that the Runtime conceptually represent the thread pool. Dropping
the runtime implies shutting it down. Thus the Runtime cannot be cloned.

What we need in Raft is a clonable handle, not a Runtime. However we still need
the runtime for shutdown. Thus the runtime is submit to a global runtime jail.
We only retrieve the runtime during shutdown, when we are sure all references
have been dropped.