Drop the Arc requirement of Persister.
Originally Arc was required because of tests. The persister implemented in
tests are shared between the Raft instance and the test, as they both need to
access stored data. However in real world, persisters are wrappers to disk
utilities, and are rarely shared. This change makes Ruaft work better in the
real world.

The requirement of tests are addressed by exposing the internal Raft persister
to tests only, as implemented in src/utils/integration_test.rs.

A forced casting has to be done in the test persister implemention.